[Game] Bug fixed: The number of rewards received for being in the game is now saved correctly: - Previously, the `timesTaken` value was not being stored in the `gave` field of the `divine_clock` table.

+++ AAEmu.Game/Core/Managers/AccountManager.cs
@@ -383,7 +383,7 @@ public class AccountManager : Singleton<AccountManager>
         }
     }
 
-    public void UpdateDivineClock(ulong accountId, uint scheduleItemId, uint timeElapsed, uint timesTaken)
+    public void UpdateDivineClock(ulong accountId, uint scheduleItemId, uint timeElapsed, sbyte timesTaken)
     {
         object accLock;
         lock (_locks)
